requirements.txt
telethon
pyarrow
azure-storage-blob
azure-identity
pandas
python-dotenv